+++ src/core/cli-command/import.command.const.ts
@@ -1,0 +1,2 @@
+export const DEFAULT_DB_PORT = '27017';
+export const DEFAULT_USER_PASSWORD = '123456';
