+++ src/core/database-client/database-client.const.ts
@@ -1,0 +1,2 @@
+export const RETRY_COUNT = 5;
+export const RETRY_TIMEOUT = 1000;
