+++ src/core/file-reader/file-reader.const.ts
@@ -1,0 +1,1 @@
+export const CHUNK_SIZE_READER = 16384;
