+++ src/core/file-writer/file-writer.const.ts
@@ -1,0 +1,1 @@
+export const CHUNK_SIZE_WRITER = 2 ** 16;
