+++ src/modules/comment/comment.const.ts
@@ -1,0 +1,5 @@
+export const MIN_LENGTH_COMMENT_TEXT = 5;
+export const MAX_LENGTH_COMMENT_TEXT = 1024;
+
+export const MIN_COMMENT_RATING = 1;
+export const MAX_COMMENT_RATING = 5;
