+++ src/modules/offer-generator/offer-generator.const.ts
@@ -1,0 +1,6 @@
+export const FIRST_WEEK_DAY = 1;
+export const LAST_WEEK_DAY = 7;
+
+export const COMMENT_COUNT = 0;
+
+export const DATE_FORMAT = 'YYYY-MM-DD';
