+++ src/modules/offer-generator/offer-generator.ts
@@ -3,8 +3,9 @@ import dayjs from 'dayjs';
 import { RentalType } from './../../types/rental-type.enum.js';
 import { generateRandomValue, getRandomBoolean, getRandomImages, getRandomItem, getRandomItems, objectToString, getCoordinates } from '../../core/helpers/index.js';
 import { MockData } from '../../types/mock-data.js';
-import { COMMENT_COUNT, DATE_FORMAT, FIRST_WEEK_DAY, LAST_WEEK_DAY, MAX_COUNT_OFFER_GUESTS, MAX_COUNT_OFFER_ROOMS, MAX_OFFER_PRICE, MAX_OFFER_RATING, MIN_COUNT_OFFER_GUESTS, MIN_COUNT_OFFER_ROOMS, MIN_OFFER_PRICE, MIN_OFFER_RATING } from '../../const.js';
 import type { OfferGeneratorInterface } from './offer-generator.interface.js';
+import { MAX_COUNT_OFFER_GUESTS, MAX_COUNT_OFFER_ROOMS, MAX_OFFER_PRICE, MAX_OFFER_RATING, MIN_COUNT_OFFER_GUESTS, MIN_COUNT_OFFER_ROOMS, MIN_OFFER_PRICE, MIN_OFFER_RATING } from '../offer/offer.const.js';
+import { COMMENT_COUNT, DATE_FORMAT, FIRST_WEEK_DAY, LAST_WEEK_DAY } from './offer-generator.const.js';
 
 export default class OfferGenerator implements OfferGeneratorInterface {
   constructor(private readonly mockData: MockData) {}
