+++ src/modules/offer/offer.const.ts
@@ -1,0 +1,17 @@
+export const MIN_OFFER_PRICE = 100;
+export const MAX_OFFER_PRICE = 100000;
+
+export const MIN_LENGTH_OFFER_TITLE = 10;
+export const MAX_LENGTH_OFFER_TITLE = 100;
+
+export const MIN_LENGTH_OFFER_DESCRIPTION = 20;
+export const MAX_LENGTH_OFFER_DESCRIPTION = 1024;
+
+export const MIN_OFFER_RATING = 1;
+export const MAX_OFFER_RATING = 5;
+
+export const MIN_COUNT_OFFER_ROOMS = 1;
+export const MAX_COUNT_OFFER_ROOMS = 8;
+
+export const MIN_COUNT_OFFER_GUESTS = 1;
+export const MAX_COUNT_OFFER_GUESTS = 10;
