+++ src/modules/user/user.const.ts
@@ -1,0 +1,5 @@
+export const MIN_LENGTH_PASSWORD = 6;
+export const MAX_LENGTH_PASSWORD = 12;
+
+export const MIN_LENGTH_USERNAME = 1;
+export const MAX_LENGTH_USERNAME = 15;
